+++ CHANGELOG.markdown
@@ -1,8 +1,13 @@
+next
+----
+* Added `Bifoldable` and `Bitraversable` instances for `Constant` from `transformers`
+
 5.2
 -----
 * Added several `Arrow`-like instances for `Tannen` so we can use it as the Cayley construction if needed.
 * Added `Data.Bifunctor.Sum`
 * Added `BifunctorFunctor`, `BifunctorMonad` and `BifunctorComonad`.
+* Backported `Bifunctor Constant` instance from `transformers`
 
 5.1
 ---